requirements.txt
elasticsearch
requests
pyyaml
requests_mock
